Refactored callee failure tests via TestWiresCalleeFailMixin.
